Optimize q_sort function
In previous version, list_concat have to go through
all elements to concat the other list.
The reason which list_concat need to go through is to find
out the tail of the list, but the method is really inefficient.
So, I add a parameter "tail" in the list_concat function so that
two lists can directly concatenate together.

The previous version, the q_insert_tail will go wrong after q_sort.
This is because q_sort function doesn't update tail to point to
the correct list element, and I think that go through the list evertime
to update tail is very inefficient. In this version, I fix it!!