Optimize list_merge function
Originally, I would like to update q->tail in sorting function, but
I think it may not be a good solution to speed up because in list_merge
function, the function needs to go through all list elements to set the
tail. Why not just set the q->tail after sorting? That's what I am doing
in this version.